# api/openapi.yaml
openapi: 3.1.0
info:
  title: video-record API
  version: 1.0.0
  description: Private movie and TV viewing diary API.
servers:
  - url: /
security:
  - sessionCookie: []
paths:
  /api/v1/setup/status:
    get:
      operationId: getSetupStatus
      summary: Read closed-initialization status
      security: []
      responses:
        '200':
          description: Setup status.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SetupStatus' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/setup/admin:
    post:
      operationId: initializeAdministrator
      summary: Create the first administrator
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CredentialsRequest' }
      responses:
        '201':
          description: Administrator created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/auth/login:
    post:
      operationId: login
      summary: Start an opaque database session
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CredentialsRequest' }
      responses:
        '200':
          description: Session created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LoginResponse' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/auth/me:
    get:
      operationId: getCurrentUser
      summary: Read the current session user
      responses:
        '200':
          description: Current user.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/auth/logout:
    post:
      operationId: logout
      summary: Revoke the current session
      parameters:
        - $ref: '#/components/parameters/CSRFToken'
      responses:
        '204': { description: Session revoked. }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/backups:
    get:
      operationId: listBackups
      summary: List consistent system backups
      responses:
        '200':
          description: Backup artifacts.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BackupList' }
        default: { $ref: '#/components/responses/Problem' }
    post:
      operationId: createBackup
      summary: Create a consistent system backup
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EmptyObject' }
      responses:
        '201':
          description: Backup created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BackupArtifact' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/backups/{filename}:
    parameters:
      - name: filename
        in: path
        required: true
        schema: { type: string, pattern: '^[^/]+\.vrbackup$' }
    get:
      operationId: downloadBackup
      summary: Download a system backup
      responses:
        '200':
          description: Backup archive.
          content:
            application/vnd.video-record.backup:
              schema: { type: string, format: binary }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/restore:
    post:
      operationId: restoreBackup
      summary: Atomically restore a system backup
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema: { $ref: '#/components/schemas/FileUpload' }
      responses:
        '200':
          description: Restore committed.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RestoreResult' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/calendar:
    get:
      operationId: getCalendar
      summary: Read the current user's viewing calendar
      parameters:
        - name: month
          in: query
          required: true
          schema: { type: string, pattern: '^\d{4}-\d{2}$' }
        - name: timezone
          in: query
          required: true
          schema: { type: string }
        - name: filter
          in: query
          required: false
          schema: { type: string, enum: [all, completed, in_progress], default: all }
      responses:
        '200':
          description: Calendar month.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CalendarResponse' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/collections:
    get:
      operationId: listCollections
      summary: List private collections
      responses:
        '200':
          description: Private collections.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CollectionList' }
        default: { $ref: '#/components/responses/Problem' }
    post:
      operationId: createCollection
      summary: Create a private collection
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateCollectionRequest' }
      responses:
        '201':
          description: Collection created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Collection' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/collections/{collectionID}/items:
    parameters:
      - name: collectionID
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      operationId: addCollectionItem
      summary: Add media to a private collection
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AddCollectionItemRequest' }
      responses:
        '204': { description: Collection item added. }
        default: { $ref: '#/components/responses/Problem' }
    put:
      operationId: replaceCollectionItems
      summary: Replace and order items in a private collection
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReplaceCollectionItemsRequest' }
      responses:
        '204': { description: Collection items replaced in the supplied order. }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/data/export:
    get:
      operationId: exportData
      summary: Export current-user viewing data
      parameters:
        - name: format
          in: query
          required: true
          schema: { type: string, enum: [json, csv] }
      responses:
        '200':
          description: Export document.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExportDocument' }
            'text/csv; charset=utf-8':
              schema: { type: string }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/data/import:
    post:
      operationId: importData
      summary: Import validated current-user viewing data
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema: { $ref: '#/components/schemas/FileUpload' }
      responses:
        '200':
          description: Import report.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ImportReport' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/integrations/accounts:
    get:
      operationId: listIntegrationAccounts
      summary: List current-user media server account metadata
      responses:
        '200':
          description: Encrypted integration account metadata without credentials or base URLs.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IntegrationAccountList' }
        default: { $ref: '#/components/responses/Problem' }
    post:
      operationId: createIntegrationAccount
      summary: Encrypt and create a current-user media server account
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateIntegrationAccountRequest' }
      responses:
        '201':
          description: Integration account created and sync jobs provisioned.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IntegrationAccount' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/integrations/accounts/{accountID}:
    parameters:
      - name: accountID
        in: path
        required: true
        schema: { type: string, format: uuid }
    delete:
      operationId: disconnectIntegrationAccount
      summary: Disconnect an owned media server account
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CSRFToken'
      responses:
        '204': { description: Integration account disconnected. }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/library:
    get:
      operationId: getLibrary
      summary: List the current user's media library
      parameters:
        - name: status
          in: query
          required: false
          schema: { $ref: '#/components/schemas/RecordStatus' }
      responses:
        '200':
          description: Cursor-shaped library page.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CursorPage' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/media/search:
    get:
      operationId: searchLocalMedia
      summary: Search the local media catalog
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string, minLength: 1 }
      responses:
        '200':
          description: Local search results.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CatalogSearchResponse' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/records/{mediaID}:
    parameters:
      - $ref: '#/components/parameters/MediaID'
    get:
      operationId: getRecord
      summary: Read the current user's media record
      responses:
        '200':
          description: Current media record.
          headers:
            ETag: { required: true, schema: { type: string } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RecordState' }
        default: { $ref: '#/components/responses/Problem' }
    put:
      operationId: updateRecord
      summary: Update a versioned current-user media record
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CSRFToken'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateRecordRequest' }
      responses:
        '200':
          description: Updated media record.
          headers:
            ETag: { required: true, schema: { type: string } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RecordState' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/records/{mediaID}/tags:
    parameters:
      - $ref: '#/components/parameters/MediaID'
    get:
      operationId: getRecordTags
      summary: Read the current user's private tags for a media record
      responses:
        '200':
          description: Current private tags.
          headers:
            ETag: { required: true, schema: { type: string } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TagsRequest' }
        default: { $ref: '#/components/responses/Problem' }
    put:
      operationId: setRecordTags
      summary: Replace private tags for a media record
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CSRFToken'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TagsRequest' }
      responses:
        '204':
          description: Tags replaced.
          headers:
            ETag: { required: true, schema: { type: string } }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/records/{mediaID}/events:
    parameters:
      - $ref: '#/components/parameters/MediaID'
    get:
      operationId: listWatchEvents
      summary: List immutable watch events for a media record
      responses:
        '200':
          description: Watch events.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WatchEventList' }
        default: { $ref: '#/components/responses/Problem' }
    post:
      operationId: createWatchEvent
      summary: Create an immutable rewatch event
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateWatchEventRequest' }
      responses:
        '201':
          description: Watch event created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WatchEvent' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/records/{mediaID}/events/{eventID}:
    parameters:
      - $ref: '#/components/parameters/MediaID'
      - name: eventID
        in: path
        required: true
        schema: { type: string, format: uuid }
    delete:
      operationId: deleteWatchEvent
      summary: Delete an owned watch event
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CSRFToken'
      responses:
        '204': { description: Watch event deleted. }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/records/{mediaID}/progress:
    parameters:
      - $ref: '#/components/parameters/MediaID'
    get:
      operationId: getEpisodeProgress
      summary: Read episode progress for a series
      responses:
        '200':
          description: Episode progress.
          headers:
            ETag: { required: true, schema: { type: string } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EpisodeProgress' }
        default: { $ref: '#/components/responses/Problem' }
    post:
      operationId: updateEpisodeProgress
      summary: Apply an idempotent episode progress action
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateEpisodeProgressRequest' }
      responses:
        '200':
          description: Updated episode progress.
          headers:
            ETag: { required: true, schema: { type: string } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EpisodeProgress' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/stats:
    get:
      operationId: getStats
      summary: Read private accessible viewing statistics
      parameters:
        - name: timezone
          in: query
          required: false
          schema: { type: string, default: UTC }
      responses:
        '200':
          description: Viewing statistics.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StatsSummary' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/sync/status:
    get:
      operationId: getSyncStatus
      summary: Read current-user integration sync status
      responses:
        '200':
          description: Sync status.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SyncStatus' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/sync/candidates:
    get:
      operationId: listSyncCandidates
      summary: List current-user sync candidates
      parameters:
        - name: status
          in: query
          required: false
          schema: { $ref: '#/components/schemas/SyncCandidateStatus' }
      responses:
        '200':
          description: Sync candidates.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SyncCandidateList' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/sync/candidates/{candidateID}/confirm:
    parameters:
      - $ref: '#/components/parameters/CandidateID'
    post:
      operationId: confirmSyncCandidate
      summary: Confirm a proposed sync match
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EmptyObject' }
      responses:
        '200':
          description: Confirmed candidate.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SyncCandidate' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/sync/candidates/{candidateID}/rematch:
    parameters:
      - $ref: '#/components/parameters/CandidateID'
    post:
      operationId: rematchSyncCandidate
      summary: Rematch a sync candidate to another local item
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RematchCandidateRequest' }
      responses:
        '200':
          description: Rematched candidate.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SyncCandidate' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/sync/candidates/{candidateID}/ignore:
    parameters:
      - $ref: '#/components/parameters/CandidateID'
    post:
      operationId: ignoreSyncCandidate
      summary: Ignore a sync candidate
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EmptyObject' }
      responses:
        '200':
          description: Ignored candidate.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SyncCandidate' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/sync/candidates/{candidateID}/custom:
    parameters:
      - $ref: '#/components/parameters/CandidateID'
    post:
      operationId: createCustomSyncCandidate
      summary: Create custom media for an unmatched sync candidate
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CustomMediaRequest' }
      responses:
        '200':
          description: Confirmed custom candidate.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SyncCandidate' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/household/events:
    get:
      operationId: listSharedEvents
      summary: List minimal shared household events
      responses:
        '200':
          description: Shared events.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SharedEventList' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/household/members:
    get:
      operationId: listHouseholdMembers
      summary: List household members
      responses:
        '200':
          description: Household members.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MemberList' }
        default: { $ref: '#/components/responses/Problem' }
    post:
      operationId: createHouseholdMember
      summary: Create a household member
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateMemberRequest' }
      responses:
        '201':
          description: Household member created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Member' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/household/participants:
    get:
      operationId: listHouseholdParticipants
      summary: List active co-viewing participants
      responses:
        '200':
          description: Active participants.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MemberList' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/household/members/{memberID}/reset-password:
    parameters:
      - $ref: '#/components/parameters/MemberID'
    post:
      operationId: resetHouseholdMemberPassword
      summary: Reset a household member password
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ResetPasswordRequest' }
      responses:
        '204': { description: Password reset and sessions revoked. }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/household/members/{memberID}/deactivate:
    parameters:
      - $ref: '#/components/parameters/MemberID'
    post:
      operationId: deactivateHouseholdMember
      summary: Deactivate a household member and sessions
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EmptyObject' }
      responses:
        '204': { description: Member deactivated. }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/household/records/{ownerID}/{mediaID}:
    parameters:
      - name: ownerID
        in: path
        required: true
        schema: { type: string, format: uuid }
      - $ref: '#/components/parameters/MediaID'
    get:
      operationId: getVisibleHouseholdRecord
      summary: Read explicitly shared household record fields
      responses:
        '200':
          description: Explicitly shared record fields.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VisibleRecord' }
        default: { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: video_record_session
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 1, maxLength: 128 }
    CSRFToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema: { type: string, minLength: 1 }
...
  responses:
    Problem:
      description: RFC 9457 Problem Details response.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
  schemas:
    EmptyObject:
      type: object
      additionalProperties: false
...
    SetupStatus:
      type: object
      additionalProperties: false
      required: [initialized, storageReady, tmdbConfigured]
      properties:
        initialized: { type: boolean }
        storageReady: { type: boolean }
        tmdbConfigured: { type: boolean }
    CredentialsRequest:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username: { type: string, minLength: 1 }
        password: { type: string, minLength: 1 }
    User:
      type: object
      additionalProperties: false
      required: [id, username, role]
      properties:
        id: { type: string, format: uuid }
        username: { type: string }
        role: { type: string, enum: [admin, member] }
    LoginResponse:
      type: object
      additionalProperties: false
      required: [user, csrfToken]
      properties:
        user: { $ref: '#/components/schemas/User' }
        csrfToken: { type: string }
...
    BackupList:
      type: array
      items: { $ref: '#/components/schemas/BackupArtifact' }
...